temp

+++ src/utils/crudUtil.ts
@@ -1,5 +1,7 @@
-import { createClient } from '@/app/supabase/supabaseClient';
+import { createClient } from '@/supabase/supabaseClient';
 import { TableInfoType, TableNameType } from '../supabase/tableInfo';
+import { User } from '@supabase/supabase-js';
+import { getUser } from './authUtil';
 
 /**
  * Supabaseに接続するクライアントインスタンス
@@ -15,9 +17,15 @@ const supabase = createClient();
  * @returns 検索結果
  */
 export const selectAll = async <T>(tableName: TableNameType): Promise<T[]> => {
+
+  // ユーザ情報の取得
+  const user: User = await getUser();
+
+  // データの取得
   const { data, error } = await supabase
     .from(tableName)
-    .select();
+    .select()
+    .eq('user_id', user.id);
 
   // ロギング
   console.log(`selectAll('${tableName}'):`, data);
@@ -38,12 +46,20 @@ export const selectAll = async <T>(tableName: TableNameType): Promise<T[]> => {
  * @returns 挿入結果
  */
 export const insertInto = async <T>(tableName: TableNameType, tableInfo: TableInfoType): Promise<T[]> => {
+
+  // ユーザ情報の取得
+  const user: User = await getUser();
+
+  // レコードにユーザIDを追加
+  tableInfo.user_id = user.id;
+
+  // データの挿入
   const { data, error } = await supabase
     .from(tableName)
     .insert([
       tableInfo,
     ])
-    .select()
+    .select();
 
   // ロギング
   console.log(`insertInto('${tableName}'):`, data);
@@ -55,16 +71,24 @@ export const insertInto = async <T>(tableName: TableNameType, tableInfo: TableIn
 };
 
 /**
+ * deleteByDate
+ * <p>
  * 日付を指定してレコードを削除する
  * 
  * @param tableName テーブル名
  * @param date 日付
  */
 export const deleteByDate = async (tableName: TableNameType, date: string) => {
+
+  // ユーザ情報の取得
+  const user: User = await getUser();
+
+  // データの削除
   const { error } = await supabase
     .from(tableName)
     .delete()
     .eq('date', date)
+    .eq('user_id', user.id);
 
   // ロギング
   if (error) {
